Parser
- moved JSON, XML, OVA and OVF parsers to separate files
